address review feedback: fix double newline, rename fs.ts, add eslint rule
- Fix double newline in writeYaml by trimming yamlStringify output before
  passing to writeToFile (which adds its own trailing newline)
- Rename fs.ts to utils.ts to avoid redundant fs/fs.ts naming
- Add ESLint rule to restrict @hyperlane-xyz/utils/fs imports in browser
  packages that use restrictedSdkAndUtilsImportRules
